ENG-4890 - Refactor Code, Add Tests
