+++ NeuroID/TrackerClassExtensions/BaseClassExtensions/CharacterSet.swift
@@ -1,0 +1,19 @@
+//
+//  CharacterSet.swift
+//  NeuroID
+//
+//  Created by Kevin Sites on 3/29/23.
+//
+
+import Foundation
+
+extension CharacterSet {
+    static let urlQueryValueAllowed: CharacterSet = {
+        let generalDelimitersToEncode = ":#[]@" // does not include "?" or "/" due to RFC 3986 - Section 3.4
+        let subDelimitersToEncode = "!$&'()*+,;="
+
+        var allowed = CharacterSet.urlQueryAllowed
+        allowed.remove(charactersIn: "\(generalDelimitersToEncode)\(subDelimitersToEncode)")
+        return allowed
+    }()
+}
