+++ NeuroID/TrackerClassExtensions/BaseClassExtensions/Data.swift
@@ -1,0 +1,34 @@
+//
+//  Data.swift
+//  NeuroID
+//
+//  Created by Kevin Sites on 3/29/23.
+//
+
+import CommonCrypto
+import Foundation
+
+extension Data {
+    public func sha256() -> String {
+        return hexStringFromData(input: digest(input: self as NSData))
+    }
+
+    private func digest(input: NSData) -> NSData {
+        let digestLength = Int(CC_SHA256_DIGEST_LENGTH)
+        var hash = [UInt8](repeating: 0, count: digestLength)
+        CC_SHA256(input.bytes, UInt32(input.length), &hash)
+        return NSData(bytes: hash, length: digestLength)
+    }
+
+    private func hexStringFromData(input: NSData) -> String {
+        var bytes = [UInt8](repeating: 0, count: input.length)
+        input.getBytes(&bytes, length: input.length)
+
+        var hexString = ""
+        for byte in bytes {
+            hexString += String(format: "%02x", UInt8(byte))
+        }
+
+        return hexString
+    }
+}
