+++ NeuroID/TrackerClassExtensions/BaseClassExtensions/Dictionary.swift
@@ -1,0 +1,19 @@
+//
+//  Dictionary.swift
+//  NeuroID
+//
+//  Created by Kevin Sites on 3/29/23.
+//
+
+import Foundation
+
+extension Dictionary {
+    func toKeyValueString() -> String {
+        return map { key, value in
+            let escapedKey = "\(key)" ?? ""
+            let escapedValue = "\(value)" ?? ""
+            return escapedKey + "=" + escapedValue
+        }
+        .joined(separator: "&")
+    }
+}
