+++ NeuroID/TrackerEvents/AppEvents.swift
@@ -1,0 +1,33 @@
+//
+//  AppEvents.swift
+//  NeuroID
+//
+//  Created by Kevin Sites on 3/29/23.
+//
+
+import Foundation
+import UIKit
+
+// MARK: - App events
+
+internal extension NeuroIDTracker {
+    func observeAppEvents() {
+        if #available(iOS 13.0, *) {
+            NotificationCenter.default.addObserver(self, selector: #selector(appMovedToBackground), name: UIScene.willDeactivateNotification, object: nil)
+
+            NotificationCenter.default.addObserver(self, selector: #selector(appMovedToForeground), name: UIScene.willEnterForegroundNotification, object: nil)
+        } else {
+            NotificationCenter.default.addObserver(self, selector: #selector(appMovedToBackground), name: UIApplication.willResignActiveNotification, object: nil)
+
+            NotificationCenter.default.addObserver(self, selector: #selector(appMovedToForeground), name: UIApplication.willEnterForegroundNotification, object: nil)
+        }
+    }
+
+    @objc func appMovedToBackground() {
+        captureEvent(event: NIDEvent(type: NIDEventName.windowBlur))
+    }
+
+    @objc func appMovedToForeground() {
+        captureEvent(event: NIDEvent(type: NIDEventName.windowFocus))
+    }
+}
