+++ SDKTest/EventTests.swift
@@ -5,11 +5,10 @@
 //  Created by Clayton Selby on 8/19/21.
 //
 
-import XCTest
 @testable import NeuroID
+import XCTest
 
 class EventTests: XCTestCase {
-    
     let clientKey = "key_live_vtotrandom_form_mobilesandbox"
     let userId = "form_mobilesandbox"
     
@@ -17,121 +16,127 @@ class EventTests: XCTestCase {
         NeuroID.configure(clientKey: clientKey)
     }
     
-    func testSetScreeName(){
+    override func setUp() {
+        // Clear out the DataStore Events after each test
+        DataStore.events = []
+    }
+    
+    func testSetScreeName() {
         NeuroID.setScreenName(screen: "WOW")
         assert(NeuroID.getScreenName() == "WOW")
     }
     
-    func testInvalidScreenName(){
+    func testInvalidScreenName() {
         NeuroID.setScreenName(screen: "WOW A SPACE")
         assert(NeuroID.getScreenName() == "WOW%20A%20SPACE")
     }
 
     func testLocale() {
         print("Locale: ", ParamsCreator.getLocale())
-        XCTAssertTrue(ParamsCreator.getLocale() != nil);
+        XCTAssertTrue(ParamsCreator.getLocale() != nil)
     }
     
     func testUserAgent() {
-        let data = ParamsCreator.getUserAgent();
+        let data = ParamsCreator.getUserAgent()
         print("User Agent: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
+
     func testSessionExpired() {
-        let data = ParamsCreator.isSessionExpired();
-        XCTAssertTrue(true);
+        let data = ParamsCreator.isSessionExpired()
+        XCTAssertTrue(true)
     }
     
-    func testSettingAndGettingAPIClientKey(){
-        let keyLookup = NeuroID.getClientKeyFromLocalStorage();
+    func testSettingAndGettingAPIClientKey() {
+        let keyLookup = NeuroID.getClientKeyFromLocalStorage()
         print("Key lookup:", keyLookup)
         XCTAssertTrue(keyLookup == clientKey)
     }
     
-    func testGetTimeStamp(){
-        let data = ParamsCreator.getTimeStamp();
+    func testGetTimeStamp() {
+        let data = ParamsCreator.getTimeStamp()
         print("Timestamp: ", data)
-        XCTAssertTrue(data != nil);
-        
+        XCTAssertTrue(data != nil)
     }
+
     func testTimeZoneOffset() {
-        let data = ParamsCreator.getTimezone();
+        let data = ParamsCreator.getTimezone()
         print("TimeZone Offset: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
     func testLanguage() {
-        let data = ParamsCreator.getLanguage();
+        let data = ParamsCreator.getLanguage()
         print("Language: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
     func testPlatform() {
-        let data = ParamsCreator.getPlatform();
+        let data = ParamsCreator.getPlatform()
         print("Platform: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
-    func testDnt(){
-        let data = ParamsCreator.getDnt();
+    func testDnt() {
+        let data = ParamsCreator.getDnt()
         print("Dnt: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
-    func testTouch(){
-        let data = ParamsCreator.getTouch();
+    func testTouch() {
+        let data = ParamsCreator.getTouch()
         print("Touch: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
-    func testGetBaseURL(){
-        let data = NeuroID.getCollectionEndpointURL();
+    func testGetBaseURL() {
+        let data = NeuroID.getCollectionEndpointURL()
         print("URL: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
-    
-    func testSDKVersion(){
+    func testSDKVersion() {
         let data = ParamsCreator.getSDKVersion()
         print("Version: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
-    func testCommandQueueNameSpace(){
+    func testCommandQueueNameSpace() {
         let data = ParamsCreator.getCommandQueueNamespace()
         print("Namespace: ", data)
-        XCTAssertTrue(data != nil);
+        XCTAssertTrue(data != nil)
     }
     
-    func testStoppingSDKIsStoppedFalse(){
-        if (NeuroID.isStopped()) {
-            NeuroID.start();
+    func testStoppingSDKIsStoppedFalse() {
+        if NeuroID.isStopped() {
+            NeuroID.start()
         }
         let isStopped = NeuroID.isStopped()
         print("Is Stopped", isStopped)
-        XCTAssertTrue(!isStopped);
+        XCTAssertTrue(!isStopped)
     }
     
-    func testStoppingSDKIsStoppedTrue(){
+    func testStoppingSDKIsStoppedTrue() {
         NeuroID.stop()
         let isStopped = NeuroID.isStopped()
         print("Is Stopped", isStopped)
-        XCTAssertTrue(isStopped);
+        XCTAssertTrue(isStopped)
     }
     
-    func testSetIUserID(){
+    func testSetIUserID() {
         NeuroID.setUserID("atestUserID")
         let params = ParamsCreator.getDefaultSessionParams()
         let uid = params["userId"] as! String
         XCTAssert(uid == "atestUserID")
     }
+
     func testEventParams() throws {
 //        let urlName = "HomeScreen"
 //        let tracker = NeuroIDTracker(userUrl: urlName)
         let urlName = "HomeScreen"
-        let testView = UIViewController();
+        let testView = UIViewController()
         let userID = "atestUserID"
-        let tracker = NeuroIDTracker(screen: urlName, controller: testView);
+        let _ = NeuroIDTracker(screen: urlName, controller: testView)
 //        let params = ParamsCreator.getDefaultSessionParams();
         NeuroID.setUserID(userID)
 //        let copyEvent = NIDEvent(type: .copy, tg: ["et": "fieldset"], x: 10, y: 10)
@@ -154,80 +159,80 @@ class EventTests: XCTestCase {
 //        XCTAssertTrue(events["y"] as! Int == 10)
     }
     
-    func testEventSubmitForm(){
+    func testEventSubmitForm() {
         let event = NeuroID.formSubmit()
         XCTAssertTrue(event.type == NIDEventName.applicationSubmit.rawValue)
     }
     
-    func testEventSubmitFormFailure(){
+    func testEventSubmitFormFailure() {
         let event = NeuroID.formSubmitFailure()
         XCTAssertTrue(event.type == NIDEventName.applicationSubmitFailure.rawValue)
     }
     
-    func testEventSubmitFormSuccess(){
+    func testEventSubmitFormSuccess() {
         let event = NeuroID.formSubmitSuccess()
         XCTAssertTrue(event.type == NIDEventName.applicationSubmitSuccess.rawValue)
     }
     
-    func testEventSetVariable(){
+    func testEventSetVariable() {
         let event = NeuroID.setCustomVariable(key: "test", v: "test2")
         XCTAssertTrue(event.type == NIDSessionEventName.setVariable.rawValue)
     }
     
-    func testLateRegistration(){
-        var view = UITextView()
+    func testLateRegistration() {
+        let view = UITextView()
         view.id = "myTextView"
-        var view2 = UITextView()
+        let view2 = UITextView()
         view2.id = "myTextView2"
         XCTAssertTrue(NeuroIDTracker.registerViewIfNotRegistered(view: view))
         
         // Ensure we dont re-add it
         XCTAssertFalse(NeuroIDTracker.registerViewIfNotRegistered(view: view))
     }
-    func testCalcSimilarity(){
+
+    func testCalcSimilarity() {
         let urlName = "HomeScreen"
-        let testView = UIViewController();
-        if (NeuroID.isStopped()) {
-            NeuroID.start();
+        let testView = UIViewController()
+        if NeuroID.isStopped() {
+            NeuroID.start()
         }
 
-        let tracker = NeuroIDTracker(screen: urlName, controller: testView);
-        var similarityLonger = tracker.calcSimilarity(previousValue: "alat", currentValue: "zlata")
+        let tracker = NeuroIDTracker(screen: urlName, controller: testView)
+        let similarityLonger = tracker.calcSimilarity(previousValue: "alat", currentValue: "zlata")
         print("Close Similarity \(similarityLonger)")
         XCTAssertTrue(similarityLonger == 0.6)
         
-        var similarityShorter = tracker.calcSimilarity(previousValue: "amuchlongerdiffere", currentValue: "wowshort")
+        let similarityShorter = tracker.calcSimilarity(previousValue: "amuchlongerdiffere", currentValue: "wowshort")
         print("Much different Similarity \(similarityShorter)")
         XCTAssertTrue(similarityLonger == 0.6)
     }
     
-    func testpercentageDifference(){
+    func testpercentageDifference() {
         let urlName = "HomeScreen"
-        let testView = UIViewController();
-        if (NeuroID.isStopped()) {
-            NeuroID.start();
+        let testView = UIViewController()
+        if NeuroID.isStopped() {
+            NeuroID.start()
         }
 
-        let tracker = NeuroIDTracker(screen: urlName, controller: testView);
-        var percentDiff = tracker.percentageDifference(newNumOrig: "20", originalNumOrig: "30")
+        let tracker = NeuroIDTracker(screen: urlName, controller: testView)
+        let percentDiff = tracker.percentageDifference(newNumOrig: "20", originalNumOrig: "30")
         print("Percentdiff \(percentDiff)")
-        
     }
     
-    func testExcludeByStringID(){
+    func testExcludeByStringID() {
         NeuroID.excludeViewByTestID(excludedView: "DontTrackMeID")
-        assert(NeuroID.excludedViewsTestIDs.contains(where: { $0 == "DontTrackMeID"}))
+        assert(NeuroID.excludedViewsTestIDs.contains(where: { $0 == "DontTrackMeID" }))
     }
     
     func testSha256Salt() {
-        var myTestString = "test"
-        var hash = myTestString.sha256()
+        let myTestString = "test"
+        let hash = myTestString.sha256()
         print("Raw hash \(hash)")
         print("Prefixed hash \(hash.prefix(8))")
         assert(!hash.prefix(8).isEmpty)
     }
     
-    func testCloseSession(){
+    func testCloseSession() {
         do {
             NeuroID.start()
             let closeSession = try NeuroID.closeSession()
@@ -238,25 +243,28 @@ class EventTests: XCTestCase {
         }
     }
      
-    func testManuallyTargetRegister(){
+    func testManuallyTargetRegister() {
         NeuroID.start()
-        var testView = UITextView();
+        let testView = UITextView()
         testView.id = "wow"
         NeuroID.manuallyRegisterTarget(view: testView)
-        var events = DataStore.getAllEvents()
-        var validEvent = events.filter {  $0.type == "REGISTER_TARGET"}
+        let events = DataStore.getAllEvents()
+        let registeredTargetEvents = events.filter { $0.type == "REGISTER_TARGET" }
+        let validEvent = registeredTargetEvents.filter { $0.tgs == "wow" }
         assert(validEvent.count == 1)
+        assert(validEvent[0].et == "UITextView::UITextView")
     }
     
-    func testRNManuallyTargetRegister(){
+    func testRNManuallyTargetRegister() {
         NeuroID.start()
-        let testView = UITextView();
+        let testView = UITextView()
         testView.id = "wow"
         let event = NeuroID.manuallyRegisterRNTarget(id: "WOW2", className: "UITextView", screenName: "HOME", placeHolder: "name here")
         let events = DataStore.getAllEvents()
-        let validEvent = events.filter {  $0.type == "REGISTER_TARGET"}
+//        let validEvent = events.filter { $0.type == "REGISTER_TARGET" }
         assert(event.etn == "INPUT")
     }
+
     func testFullPayload() {
         NeuroID.start()
         var tracker: NeuroIDTracker?
@@ -300,7 +308,7 @@ class EventTests: XCTestCase {
         textfield.text = "text"
         let lengthValue = "S~C~~\(textfield.text?.count ?? 0)"
         let hashValue = textfield.text?.sha256().prefix(8).string
-        let inputTG = ParamsCreator.getTGParamsForInput(eventName: NIDEventName.input, view: textfield, type: "text", attrParams: ["v": lengthValue, "hash": textfield.text])
+        let inputTG = ParamsCreator.getTGParamsForInput(eventName: NIDEventName.input, view: textfield, type: "text", attrParams: ["v": lengthValue, "hash": textfield.text ?? "emptyHash"])
         var inputEvent = NIDEvent(type: NIDEventName.input, tg: inputTG)
         inputEvent.v = lengthValue
         inputEvent.hv = hashValue
@@ -311,12 +319,12 @@ class EventTests: XCTestCase {
         textfield.text = "text_match"
         let sm = tracker?.calcSimilarity(previousValue: "text", currentValue: "text_match") ?? 0
         let pd = tracker?.percentageDifference(newNumOrig: "text", originalNumOrig: "text_match") ?? 0
-        let textChangeTG = ParamsCreator.getTGParamsForInput(eventName: NIDEventName.textChange, view: textfield, type: "text", attrParams: ["v": lengthValue, "hash": textfield.text])
-        var textChangeEvent = NIDEvent(type:NIDEventName.textChange, tg: textChangeTG, sm: sm, pd: pd)
+        let textChangeTG = ParamsCreator.getTGParamsForInput(eventName: NIDEventName.textChange, view: textfield, type: "text", attrParams: ["v": lengthValue, "hash": textfield.text ?? "emptyHash"])
+        var textChangeEvent = NIDEvent(type: NIDEventName.textChange, tg: textChangeTG, sm: sm, pd: pd)
         textChangeEvent.v = lengthValue
         var shaText = textfield.text ?? ""
-        if (shaText != "") {
-         shaText = shaText.sha256().prefix(8).string
+        if shaText != "" {
+            shaText = shaText.sha256().prefix(8).string
         }
         textChangeEvent.hv = shaText
         textChangeEvent.tgs = TargetValue.string(textfield.id).toString()
@@ -338,13 +346,13 @@ class EventTests: XCTestCase {
         let randomString = UUID().uuidString
         let pageid = randomString.replacingOccurrences(of: "-", with: "").prefix(12)
         
-        let neuroHTTPRequest = NeuroHTTPRequest.init(clientId: ParamsCreator.getClientId(), environment: NeuroID.getEnvironment(), sdkVersion: ParamsCreator.getSDKVersion(), pageTag: NeuroID.getScreenName() ?? "UNKNOWN", responseId: ParamsCreator.generateUniqueHexId(), siteId: "_", userId: ParamsCreator.getUserID() ?? "", jsonEvents: events, tabId: "\(tabId)", pageId: "\(pageid)", url: "ios://\(NeuroID.getScreenName() ?? "")")
+        let neuroHTTPRequest = NeuroHTTPRequest(clientId: ParamsCreator.getClientId(), environment: NeuroID.getEnvironment(), sdkVersion: ParamsCreator.getSDKVersion(), pageTag: NeuroID.getScreenName() ?? "UNKNOWN", responseId: ParamsCreator.generateUniqueHexId(), siteId: "_", userId: ParamsCreator.getUserID() ?? "", jsonEvents: events, tabId: "\(tabId)", pageId: "\(pageid)", url: "ios://\(NeuroID.getScreenName() ?? "")")
         /// Transform event into json
         do {
             let encoder = JSONEncoder()
             let values = try encoder.encode(neuroHTTPRequest)
             let str = String(data: values, encoding: .utf8)
-            print(str)
+            print(str as Any)
             let filename = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)[0].appendingPathComponent("payload.txt")
             print("************\(filename)*************")
             try str?.write(to: filename, atomically: true, encoding: String.Encoding.utf8)
@@ -352,5 +360,91 @@ class EventTests: XCTestCase {
             print(error)
         }
         assert(neuroHTTPRequest != nil)
+    }
+    
+    // Specific UI Class Registration Tests
+    func testUITextFieldRegistration() {
+        NeuroID.start()
+        let testView = UITextField()
+        testView.id = "UITextField"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 1)
+        assert(filteredRegisteredTargets[0].et == "UITextField::UITextField")
+    }
+    
+    func testUITextViewRegistration() {
+        NeuroID.start()
+        let testView = UITextView()
+        testView.id = "UITextView"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 1)
+        assert(filteredRegisteredTargets[0].et == "UITextView::UITextView")
+    }
+    
+    func testUIButtonRegistration() {
+        NeuroID.start()
+        let testView = UIButton()
+        testView.id = "UIButton"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 1)
+        assert(filteredRegisteredTargets[0].et == "UIButton::UIButton")
+    }
+    
+    func testUIDatePickerRegistration() {
+        NeuroID.start()
+        let testView = UIDatePicker()
+        testView.id = "UIDatePicker"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 1)
+        assert(filteredRegisteredTargets[0].et == "UIDatePicker::UIDatePicker")
+    }
+    
+    // UI Class Registrations that are NOT implemented
+    func testUISliderNotRegistered() {
+        NeuroID.start()
+        let testView = UISlider()
+        testView.id = "UISlider"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 0)
+    }
+    
+    func testUISwitchNotRegistered() {
+        NeuroID.start()
+        let testView = UISwitch()
+        testView.id = "UISwitch"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 0)
+    }
+    
+    func testUITableViewCellNotRegistered() {
+        NeuroID.start()
+        let testView = UITableViewCell()
+        testView.id = "UITableViewCell"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 0)
+    }
+    
+    func testUIPickerViewNotRegistered() {
+        NeuroID.start()
+        let testView = UIPickerView()
+        testView.id = "UIPickerView"
+        NeuroID.manuallyRegisterTarget(view: testView)
+        let events = DataStore.getAllEvents()
+        let filteredRegisteredTargets = events.filter { $0.type == "REGISTER_TARGET" }
+        assert(filteredRegisteredTargets.count == 0)
     }
 }
